Added LED support for BBB and ATmega
LED program which lights up a LED connected to pin 13 when the ATmega is
initialized. The LED starts to flash when packets are sent via i2c. It
switch on/off for every 10th sent package.

The LED program for the BBB hav the same functionality. The LED lights
up when the hugin program is initialized and it starts flashing for
every 10th package sent to the simulink program. The LED is wired to
GPIO30 on the BBB - pin 9.11